Separate real estate lab server exports
