+++ lib/real-estate-lab/index.ts
@@ -1,3 +1,1 @@
 export * from "./contracts";
-export { realEstateLabContentAdapter } from "./adapters/realEstateLabContentAdapter";
-export { realEstateLabItemAdapter } from "./adapters/realEstateLabItemAdapter";
