+++ lib/real-estate-lab/server.ts
@@ -1,0 +1,4 @@
+import "server-only";
+
+export { realEstateLabContentAdapter } from "./adapters/realEstateLabContentAdapter";
+export { realEstateLabItemAdapter } from "./adapters/realEstateLabItemAdapter";
